Renamed magic.py to ipymagic.py.
