+++ snakeviz/__init__.py
@@ -1,2 +1,2 @@
 from .version import __version__
-from .magic import *
+from .ipymagic import *
